feature: Adicionado Prisma ORM para interagir com o banco de dados e configurado no projeto
This commit introduces Prisma as the ORM for database interactions within the project. It includes the installation of `@prisma/client` and `prisma` as dependencies, along with necessary configurations. Additionally, the `.gitignore` file has been updated to exclude the generated Prisma client directory, preventing it from being tracked in the repository.